feat: exercice has niveau CRUD : check + delete

+++ src/app/datamapper/exerciceHasNiveau.ts
@@ -46,12 +46,43 @@ class ExerciceHasNiveauDataMapper extends CoreDataMapper {
         }
     }
 
+    
+    //& Validate niveau by user
+    async checkNiveau(exerciceId: number, userId: number) {
+        if(this.client instanceof pg.Pool) {
+            const preparedQuery = {
+                text: `SELECT niveau_id
+                FROM "${this.tableName}" 
+                WHERE exercice_id = $1
+                AND user_id = $2
+                ;`,
+                values: [exerciceId, userId]
+            }
+            const result = await this.client.query(preparedQuery);
+            return result.rows;
+        }
+    }
+
     //& Delete after validation
     async deleteValidatedNiveau( exerciceId: number, userId: number) {
         if(this.client instanceof pg.Pool) {
             const preparedQuery = {
                 text: `DELETE FROM "${this.tableName}" WHERE "validated" = 'true'
                 AND "exercice_id" = $1
+                AND "user_id" = $2
+                ;`,
+                values: [exerciceId, userId]
+            }
+            const result = await this.client.query(preparedQuery);
+            return result.rows;
+        }
+    }
+
+    async deleteExerciceHasNiveau( exerciceId: number, userId: number) {
+        if(this.client instanceof pg.Pool) {
+            const preparedQuery = {
+                text: `DELETE FROM "${this.tableName}" WHERE 
+                "exercice_id" = $1
                 AND "user_id" = $2
                 ;`,
                 values: [exerciceId, userId]
